[taotao-cloud-project-1631] update[ admin to monitor]

+++ taotao-cloud-microservice/taotao-cloud-message/taotao-cloud-message-biz/src/main/java/com/taotao/cloud/message/biz/austin/web/service/impl/DataServiceImpl.java
@@ -172,12 +172,12 @@ public class DataServiceImpl implements DataService {
 					itemsVO.setBusinessId(String.valueOf(smsRecord.getMessageTemplateId()));
 					itemsVO.setContent(smsRecord.getMsgContent());
 					itemsVO.setSendType(SmsStatus.getDescriptionByStatus(smsRecord.getStatus()));
-					itemsVO.setSendTime(DateUtil.format(new Date(Long.valueOf(smsRecord.getCreated() * 1000L)), DatePattern.NORM_DATETIME_PATTERN));
+					itemsVO.setSendTime(DateUtil.format(new Date(smsRecord.getCreated() * 1000L), DatePattern.NORM_DATETIME_PATTERN));
 
 				} else {
 					itemsVO.setReceiveType(SmsStatus.getDescriptionByStatus(smsRecord.getStatus()));
 					itemsVO.setReceiveContent(smsRecord.getReportContent());
-					itemsVO.setReceiveTime(DateUtil.format(new Date(Long.valueOf(smsRecord.getUpdated() * 1000L)), DatePattern.NORM_DATETIME_PATTERN));
+					itemsVO.setReceiveTime(DateUtil.format(new Date(smsRecord.getUpdated() * 1000L), DatePattern.NORM_DATETIME_PATTERN));
 				}
 			}
 			itemsVOS.add(itemsVO);
